Update ring.html
Due all respect i have for your awesome test bed i have made some modifications so that it alligns with my research on stop and go traffic waves and paths to its mitigaion The ring.html has been enhanced some of the extra features have been deleted just to make space for the purpose specific usage (RL-modeling for vehicle IDs and other function buttons in the bottom segment of the page) each of details and functions will be explained further with more detail below. changes made:

an ID section has been added in the top right corner that indicated the ids of the vehicles
now the record data button records the movement of one veh by its id instead of all of them, you could also write in the id of the vehicle but if you leave it empty it would do veh 200 by default this way you can make better analysis on driving behaviors specially if you are implementing a costumized model like me.
there are bunch of other buttons which self explain. like the local RL model and its size and gap between veh 200 and 222 (the one in front of it)
controller type when you are using a hybrid model
laps for counting during recording data so you have track of how many laps have been taken by veh 200 (leading veh)