Update ring.js
In this section i have made a costumized self trained unsupervized RL model for veh id 200 that updates itself on each run and is stored locally in the browser. also a lot of research alligned costumized enhancements have been done that have been explained in detail in the upper directories ring.html running the code is highly suggest to visually see the changes made. most of the buttons functionality have been declared here.
note that these are suggestions and by no mean am i requesting you to change the original code. the changes as mentioned earlier were research alligned for creating a controller that mitigates stop and go waves